feat(rit-tests-v8): migrate reaxt-18-tests-v8 to generic app which uses RIT
