--- apps/react-18-tests-v8/.eslintrc.json
@@ -1,9 +1,0 @@
-{
-  "extends": ["plugin:@fluentui/eslint-plugin/react--legacy"],
-  "root": true,
-  "rules": {
-    "no-restricted-globals": "off",
-    "import/no-extraneous-dependencies": "off",
-    "@fluentui/max-len": "off"
-  }
-}
