--- apps/react-18-tests-v8/CHANGELOG.md
@@ -1,86 +1,0 @@
-# Change Log - @fluentui/react-18-tests-v8
-
-This log was last generated on Thu, 22 Sep 2022 07:43:45 GMT and should not be manually modified.
-
-<!-- Start content -->
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Thu, 22 Sep 2022 07:43:45 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.96.1 ([PR #24871](https://github.com/microsoft/fluentui/pull/24871) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Mon, 19 Sep 2022 07:47:30 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.96.0 ([PR #24771](https://github.com/microsoft/fluentui/pull/24771) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Fri, 16 Sep 2022 07:37:32 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.95.1 ([PR #24822](https://github.com/microsoft/fluentui/pull/24822) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Thu, 15 Sep 2022 19:15:27 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Minor changes
-
-- Bump @fluentui/react to v8.95.0 ([PR #24792](https://github.com/microsoft/fluentui/pull/24792) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Tue, 13 Sep 2022 07:41:03 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.94.4 ([PR #24615](https://github.com/microsoft/fluentui/pull/24615) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Thu, 08 Sep 2022 20:52:16 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.94.3 ([PR #24701](https://github.com/microsoft/fluentui/pull/24701) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Wed, 07 Sep 2022 07:54:43 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.94.2 ([PR #24663](https://github.com/microsoft/fluentui/pull/24663) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Mon, 05 Sep 2022 07:38:30 GMT 
-[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-18-tests-v8_v1.0.0..@fluentui/react-18-tests-v8_v1.0.0)
-
-### Patches
-
-- Bump @fluentui/react to v8.94.1 ([PR #24372](https://github.com/microsoft/fluentui/pull/24372) by beachball)
-
-## [1.0.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-18-tests-v8_v1.0.0)
-
-Fri, 02 Sep 2022 07:48:53 GMT
-
-### Patches
-
-- Bump @fluentui/react to v8.94.0 ([PR #24394](https://github.com/microsoft/fluentui/pull/24394) by beachball)
-- Bump @fluentui/react-hooks to v8.6.11 ([PR #24394](https://github.com/microsoft/fluentui/pull/24394) by beachball)
