--- apps/react-18-tests-v8/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@fluentui/react-18-tests-v8",
-  "description": "React 18 test application",
-  "version": "8.0.0",
-  "private": true,
-  "scripts": {
-    "type-check": "just-scripts type-check",
-    "type-check:integration": "tsc -p tsconfig.react-18.json --noEmit",
-    "format": "prettier -w . --ignore-path ../.prettierignore",
-    "format:check": "yarn format -c",
-    "e2e": "cypress run --component",
-    "e2e:local": "cypress open --component",
-    "e2e:integration": "cypress run --component --config-file cypress-react-18.config.ts",
-    "lint": "ESLINT_USE_FLAT_CONFIG=false eslint --ext .js,.ts,.tsx ./src",
-    "test": "jest --passWithNoTests",
-    "test:integration": "jest --passWithNoTests -u -c jest-react-18.config.js",
-    "start": "webpack-dev-server --mode=development"
-  },
-  "devDependencies": {
-    "@fluentui/eslint-plugin": "*",
-    "@fluentui/scripts-cypress": "*",
-    "@fluentui/scripts-jest": "*",
-    "@fluentui/scripts-tasks": "*",
-    "@fluentui/scripts-webpack": "*",
-    "@fluentui/scripts-storybook": "*"
-  },
-  "dependencies": {
-    "@cypress/react": "9.0.1",
-    "@fluentui/react": "*",
-    "@fluentui/react-cards": "*",
-    "@fluentui/react-hooks": "*",
-    "@fluentui/react-examples": "*",
-    "@fluentui/font-icons-mdl2": "*",
-    "@testing-library/dom": "^10.1.0",
-    "@testing-library/react": "^16.0.0",
-    "@types/react": "18.3.20",
-    "@types/react-dom": "18.3.6",
-    "cypress-real-events": "1.14.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-test-renderer": "18.3.1",
-    "react-is": "18.3.1",
-    "tslib": "^2.1.0"
-  }
-}
