--- apps/react-18-tests-v8/project.json
@@ -1,31 +1,0 @@
-{
-  "name": "react-18-tests-v8",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "application",
-  "tags": ["v8"],
-  "implicitDependencies": [],
-  "targets": {
-    "test": {
-      "dependsOn": ["^build"]
-    },
-    "type-check:integration": {
-      "dependsOn": ["^build"],
-      "inputs": [
-        "{projectRoot}/tsconfig.react-18.json",
-        "{workspaceRoot}/packages/react-examples/*/stories/**/index.stories.tsx"
-      ]
-    },
-    "e2e:integration": {
-      "dependsOn": [],
-      "inputs": ["{projectRoot}/cypress-react-18.config.ts", "{workspaceRoot}/packages/react-examples/**/*.e2e.ts(x)?"]
-    },
-    "test:integration": {
-      "dependsOn": ["^build"],
-      "inputs": ["{projectRoot}/jest-react-18.config.js", "{workspaceRoot}/packages/react/**/*.(test|spec).ts(x)?"]
-    },
-    "build:integration": {
-      "dependsOn": ["^build"],
-      "inputs": ["{workspaceRoot}/packages/react/**/*.(test|spec).ts(x)?"]
-    }
-  }
-}
