--- apps/react-18-tests-v8/src/index.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { createRoot } from 'react-dom/client';
-import { App } from './App';
-
-const root = createRoot(document.getElementById('root') as HTMLElement);
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-);
