--- apps/react-18-tests-v8/tsconfig.app.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "lib": ["ES2019", "dom"],
-    "outDir": "dist",
-    "declaration": true,
-    "declarationDir": "dist/types",
-    "inlineSources": true,
-    "types": []
-  },
-  "exclude": ["**/*.spec.ts", "**/*.spec.tsx", "**/*.test.ts", "**/*.test.tsx", "**/*.cy.ts", "**/*.cy.tsx"],
-  "include": ["./src/**/*.ts", "./src/**/*.tsx"]
-}
