--- apps/react-18-tests-v8/tsconfig.cy.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "jsx": "react",
-    "isolatedModules": false,
-    "types": [],
-    "lib": ["ES2019", "dom"]
-  },
-  "files": [
-    "../../node_modules/@types/node/index.d.ts",
-    "../../node_modules/@cypress/react/dist/index.d.ts",
-    "../../node_modules/cypress-real-events/index.d.ts"
-  ],
-  "include": ["./src/**/*.cy.ts", "./src/**/*.cy.tsx"]
-}
