--- apps/react-18-tests-v8/tsconfig.react-18.json
@@ -1,34 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2019",
-    "module": "esnext",
-    "moduleResolution": "node",
-    "lib": ["ES2019", "dom"],
-    "sourceMap": true,
-    "strict": true,
-    "strictFunctionTypes": false,
-    "strictPropertyInitialization": false,
-    "pretty": true,
-    "noEmit": true,
-    "isolatedModules": true,
-    "importHelpers": true,
-    "jsx": "react",
-    "noUnusedLocals": true,
-    "preserveConstEnums": true,
-    "skipLibCheck": true,
-    "typeRoots": ["./node_modules/@types"],
-    "types": [],
-    "baseUrl": ".",
-    "paths": {
-      "react/jsx-runtime": ["./node_modules/@types/react/jsx-runtime.d.ts"],
-      "react": ["./node_modules/@types/react/index.d.ts"],
-      "react-dom": ["./node_modules/@types/react-dom/index.d.ts"]
-    }
-  },
-  "exclude": [
-    "../../packages/react-examples/src/react-charting/**",
-    "../../packages/react-examples/src/react-experiments/**"
-  ],
-  "include": ["../../packages/react-examples/**/index.stories.tsx"],
-  "files": ["../../typings/static-assets/index.d.ts"]
-}
