--- apps/react-18-tests-v8/tsconfig.spec.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "types": []
-  },
-  "files": ["../../node_modules/@types/jest/index.d.ts", "../../node_modules/@types/node/index.d.ts"],
-  "include": ["**/*.spec.ts", "**/*.spec.tsx", "**/*.test.ts", "**/*.test.tsx", "**/*.d.ts"]
-}
