+++ apps/rit-tests-v8/.eslintrc.json
@@ -1,0 +1,14 @@
+{
+  "extends": ["plugin:@fluentui/eslint-plugin/react--legacy"],
+  "root": true,
+  "ignorePatterns": ["!**/*"],
+  "rules": {
+    "@nx/workspace-no-restricted-globals": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off",
+    "import/no-extraneous-dependencies": "off",
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/naming-convention": "off",
+    "@typescript-eslint/no-restricted-types": "off",
+    "no-console": "off"
+  }
+}
