+++ apps/rit-tests-v8/project.json
@@ -1,0 +1,229 @@
+{
+  "name": "rit-tests-v8",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "application",
+  "implicitDependencies": [],
+  "tags": ["vNext"],
+  "metadata": {
+    "targetGroups": {
+      "React Integration Tester": [
+        "test-rit--17--prepare",
+        "test-rit--17--test",
+        "test-rit--17--e2e",
+        "test-rit--19--prepare",
+        "test-rit--19--test",
+        "test-rit--19--e2e",
+        "test-rit"
+      ]
+    }
+  },
+  "targets": {
+    "test-rit": {
+      "executor": "nx:noop",
+      "cache": true,
+      "dependsOn": [
+        {
+          "target": "test-rit--17--test",
+          "projects": "self",
+          "params": "forward"
+        },
+        {
+          "target": "test-rit--17--e2e",
+          "projects": "self",
+          "params": "forward"
+        },
+        {
+          "target": "test-rit--19--test",
+          "projects": "self",
+          "params": "forward"
+        },
+        {
+          "target": "test-rit--19--e2e",
+          "projects": "self",
+          "params": "forward"
+        }
+      ],
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": [],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 17, 19",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "configurations": {},
+      "options": {},
+      "parallelism": true
+    },
+    "test-rit--17--prepare": {
+      "options": {
+        "cwd": "{projectRoot}",
+        "command": "yarn rit --prepare-only --no-install --project-id ci --react 17 --verbose"
+      },
+      "cache": true,
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": ["{workspaceRoot}/tmp/rit/react-17/rit-tests-v8-ci"],
+      "dependsOn": ["^build"],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 17",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "executor": "nx:run-commands",
+      "configurations": {},
+      "parallelism": true
+    },
+    "test-rit--17--test": {
+      "options": {
+        "cwd": "{projectRoot}",
+        "command": "yarn rit --project-id ci --react 17 --run test --verbose"
+      },
+      "cache": true,
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": [],
+      "dependsOn": ["test-rit--17--prepare"],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 17",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "executor": "nx:run-commands",
+      "configurations": {},
+      "parallelism": true
+    },
+    "test-rit--17--e2e": {
+      "options": {
+        "cwd": "{projectRoot}",
+        "command": "yarn rit --project-id ci --react 17 --run e2e --verbose"
+      },
+      "cache": true,
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": [],
+      "dependsOn": ["test-rit--17--prepare"],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 17",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "executor": "nx:run-commands",
+      "configurations": {},
+      "parallelism": true
+    },
+    "test-rit--19--prepare": {
+      "options": {
+        "cwd": "{projectRoot}",
+        "command": "yarn rit --prepare-only --no-install --project-id ci --react 19 --verbose"
+      },
+      "cache": true,
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": ["{workspaceRoot}/tmp/rit/react-19/rit-tests-v8-ci"],
+      "dependsOn": ["^build"],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 19",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "executor": "nx:run-commands",
+      "configurations": {},
+      "parallelism": true
+    },
+    "test-rit--19--test": {
+      "options": {
+        "cwd": "{projectRoot}",
+        "command": "yarn rit --project-id ci --react 19 --run test --verbose"
+      },
+      "cache": true,
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": [],
+      "dependsOn": ["test-rit--19--prepare"],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 19",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "executor": "nx:run-commands",
+      "configurations": {},
+      "parallelism": true
+    },
+    "test-rit--19--e2e": {
+      "options": {
+        "cwd": "{projectRoot}",
+        "command": "yarn rit --project-id ci --react 19 --run e2e --verbose"
+      },
+      "cache": true,
+      "inputs": [
+        "default",
+        "production",
+        "^production",
+        "{workspaceRoot}/jest.preset.js",
+        "{workspaceRoot}/tools/react-integration-testing/**"
+      ],
+      "outputs": [],
+      "dependsOn": ["test-rit--19--prepare"],
+      "metadata": {
+        "technologies": ["react-integration-tester"],
+        "description": "Run react integration tests against React 19",
+        "help": {
+          "command": "yarn rit --help",
+          "example": {}
+        }
+      },
+      "executor": "nx:run-commands",
+      "configurations": {},
+      "parallelism": true
+    }
+  }
+}
