+++ apps/rit-tests-v8/rit.config.js
@@ -1,0 +1,18 @@
+// @ts-check
+
+/** @type {import('@fluentui/react-integration-tester').Config} */
+const config = {
+  react: {
+    17: {
+      runConfig: { 'type-check': { configPath: 'tsconfig.r17.json' } },
+    },
+    18: {
+      runConfig: { 'type-check': { configPath: 'tsconfig.r18.json' } },
+    },
+    19: {
+      runConfig: { 'type-check': { configPath: 'tsconfig.r19.json' } },
+    },
+  },
+};
+
+module.exports = config;
