+++ apps/rit-tests-v8/src/react-18/dts-issues.tsx
@@ -1,0 +1,1 @@
+export {};
