+++ apps/rit-tests-v8/src/shared/dts-issues.tsx
@@ -1,0 +1,1 @@
+export {};
