+++ apps/rit-tests-v8/tsconfig.r17.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.lib.json",
+  "include": ["./src/shared/", "./src/react-17/**/*.ts", "./src/react-17/**/*.tsx"]
+}
